Added the interface for the Simulation.run method
Refactored some variables and added a set of documentation to the new run method to describe how the method should work

Refactored

Restructured the validation functions for Production.sample

Added age values as part of the Production.sample method

Added total surface area as an attribute to the Surface object

Improved the argument validation functions and refactored some of the variables to clarify their meaning.

Added the populate method to run.

Added ability to save intermediate simulation results to file

fixed typo

added ability to save frames to vtk file during run

set ignore_time to false otherwise I need a variable called xtime on the mesh

Improved the handling of time series data in simulations.

renabled production test plots

Fixed tests with new time variable output

Fixed the issue that was causing the wrong age value to be saved in the initial step of a run. Also don't run the emplace code if there are no impacts to emplace.

Added a safety check to prevent the age computation from running if there are no craters returned by the sampler

Fixed lingering issues with time variables and edge cases with 0 craters forming