Updated documentation
Fixed some issues with the documenation that was preventing local building. Added in the LocalSurface class to the API docs.

Began a shift to put the hillshade plotting tool and the raster output tool to work on local surfaces

Expanded LocalSurface nodes to include all associated with buffer faces, so that they can be exported without causing a failure.

Adjusted projection values. Still not quite right.